Replace int -> long arrays

+++ homework/episode1/src/main/java/edu/technopolis/homework/MatrixMultiplication.java
@@ -20,13 +20,13 @@ public class MatrixMultiplication {
 
     //******************************************************************************************
 
-    public static int[][] multiply(int[][] a, int[][] b) {
+    public static long[][] multiply(long[][] a, long[][] b) {
 
         int rowsA = a.length;
         int columnsB = b[0].length;
         int columnsA_rowsB = a[0].length;
 
-        int[][] c = new int[rowsA][columnsB];
+        long[][] c = new long[rowsA][columnsB];
 
         for (int i = 0; i < rowsA; i++) {
             for (int j = 0; j < columnsB; j++) {
@@ -42,14 +42,14 @@ public class MatrixMultiplication {
 
     //******************************************************************************************
 
-    public static int[][] multiplyTransposed(int[][] a, int[][] b) {
+    public static long[][] multiplyTransposed(long[][] a, long[][] b) {
 
         int rowsA = a.length;
         int columnsB = b[0].length;
         int columnsA_rowsB = a[0].length;
 
-        int columnB[] = new int[columnsA_rowsB];
-        int[][] c = new int[rowsA][columnsB];
+        long columnB[] = new long[columnsA_rowsB];
+        long[][] c = new long[rowsA][columnsB];
 
 
         for (int j = 0; j < columnsB; j++) {
@@ -72,11 +72,11 @@ public class MatrixMultiplication {
 
     //******************************************************************************************
 
-    private static int[][] summation(int[][] a, int[][] b) {
+    private static long[][] summation(long[][] a, long[][] b) {
 
         int n = a.length;
         int m = a[0].length;
-        int[][] c = new int[n][m];
+        long[][] c = new long[n][m];
 
         for (int i = 0; i < n; i++) {
             for (int j = 0; j < m; j++) {
@@ -88,11 +88,11 @@ public class MatrixMultiplication {
 
     //******************************************************************************************
 
-    private static int[][] subtraction(int[][] a, int[][] b) {
+    private static long[][] subtraction(long[][] a, long[][] b) {
 
         int n = a.length;
         int m = a[0].length;
-        int[][] c = new int[n][m];
+        long[][] c = new int[n][m];
 
         for (int i = 0; i < n; i++) {
             for (int j = 0; j < m; j++) {
@@ -104,9 +104,9 @@ public class MatrixMultiplication {
 
     //******************************************************************************************
 
-    private static int[][] addition2SquareMatrix(int[][] a, int n) {
+    private static long[][] addition2SquareMatrix(long[][] a, int n) {
 
-        int[][] result = new int[n][n];
+        long[][] result = new long[n][n];
 
         for (int i = 0; i < a.length; i++) {
             System.arraycopy(a[i], 0, result[i], 0, a[i].length);
@@ -116,8 +116,8 @@ public class MatrixMultiplication {
 
     //******************************************************************************************
 
-    private static int[][] getSubmatrix(int[][] a, int n, int m) {
-        int[][] result = new int[n][m];
+    private static long[][] getSubmatrix(long[][] a, int n, int m) {
+        long[][] result = new long[n][m];
         for (int i = 0; i < n; i++) {
             System.arraycopy(a[i], 0, result[i], 0, m);
         }
@@ -126,7 +126,7 @@ public class MatrixMultiplication {
 
     //******************************************************************************************
 
-    private static void splitMatrix(int[][] a, int[][] a11, int[][] a12, int[][] a21, int[][] a22) {
+    private static void splitMatrix(long[][] a, long[][] a11, long[][] a12, long[][] a21, long[][] a22) {
 
         int n = a.length >> 1;
 
@@ -140,10 +140,10 @@ public class MatrixMultiplication {
 
     //******************************************************************************************
 
-    private static int[][] collectMatrix(int[][] a11, int[][] a12, int[][] a21, int[][] a22) {
+    private static long[][] collectMatrix(long[][] a11, long[][] a12, long[][] a21, long[][] a22) {
 
         int n = a11.length;
-        int[][] a = new int[n << 1][n << 1];
+        long[][] a = new long[n << 1][n << 1];
 
         for (int i = 0; i < n; i++) {
             System.arraycopy(a11[i], 0, a[i], 0, n);
@@ -161,40 +161,40 @@ public class MatrixMultiplication {
      * Multi-threaded matrix multiplication
      * algorithm by Strassen
      */
-    private static class myRecursiveTask extends RecursiveTask<int[][]> {
+    private static class myRecursiveTask extends RecursiveTask<long[][]> {
         private static final long serialVersionUID = -433764214304695286L;
 
         int n;
-        int[][] a;
-        int[][] b;
+        long[][] a;
+        long[][] b;
 
-        public myRecursiveTask(int[][] a, int[][] b, int n) {
+        public myRecursiveTask(long[][] a, long[][] b, int n) {
             this.a = a;
             this.b = b;
             this.n = n;
         }
 
         /**
-         * @return the integer matrix by
+         * @return the long matrix by
          * multiplying 2 matrices at each other
          */
         @Override
-        protected int[][] compute() {
+        protected long[][] compute() {
             if (n <= 128) {
                 return multiplyTransposed(a, b);
             }
 
             n >>= 1;
 
-            int[][] a11 = new int[n][n];
-            int[][] a12 = new int[n][n];
-            int[][] a21 = new int[n][n];
-            int[][] a22 = new int[n][n];
+            long[][] a11 = new long[n][n];
+            long[][] a12 = new long[n][n];
+            long[][] a21 = new long[n][n];
+            long[][] a22 = new long[n][n];
 
-            int[][] b11 = new int[n][n];
-            int[][] b12 = new int[n][n];
-            int[][] b21 = new int[n][n];
-            int[][] b22 = new int[n][n];
+            long[][] b11 = new long[n][n];
+            long[][] b12 = new long[n][n];
+            long[][] b21 = new long[n][n];
+            long[][] b22 = new long[n][n];
 
             splitMatrix(a, a11, a12, a21, a22);
             splitMatrix(b, b11, b12, b21, b22);
@@ -215,18 +215,18 @@ public class MatrixMultiplication {
             task_p6.fork();
             task_p7.fork();
 
-            int[][] p1 = task_p1.join();
-            int[][] p2 = task_p2.join();
-            int[][] p3 = task_p3.join();
-            int[][] p4 = task_p4.join();
-            int[][] p5 = task_p5.join();
-            int[][] p6 = task_p6.join();
-            int[][] p7 = task_p7.join();
+            long[][] p1 = task_p1.join();
+            long[][] p2 = task_p2.join();
+            long[][] p3 = task_p3.join();
+            long[][] p4 = task_p4.join();
+            long[][] p5 = task_p5.join();
+            long[][] p6 = task_p6.join();
+            long[][] p7 = task_p7.join();
 
-            int[][] c11 = summation(summation(p1, p4), subtraction(p7, p5));
-            int[][] c12 = summation(p3, p5);
-            int[][] c21 = summation(p2, p4);
-            int[][] c22 = summation(subtraction(p1, p2), summation(p3, p6));
+            long[][] c11 = summation(summation(p1, p4), subtraction(p7, p5));
+            long[][] c12 = summation(p3, p5);
+            long[][] c21 = summation(p2, p4);
+            long[][] c22 = summation(subtraction(p1, p2), summation(p3, p6));
 
             return collectMatrix(c11, c12, c21, c22);
         }
@@ -235,15 +235,15 @@ public class MatrixMultiplication {
 
     //******************************************************************************************
 
-    public static int[][] multiStrassenForkJoin(int[][] a, int[][] b) {
+    public static long[][] multiStrassenForkJoin(long[][] a, long[][] b) {
 
         int nn = getNewDimension(a, b);
-        int[][] a_n = addition2SquareMatrix(a, nn);
-        int[][] b_n = addition2SquareMatrix(b, nn);
+        long[][] a_n = addition2SquareMatrix(a, nn);
+        long[][] b_n = addition2SquareMatrix(b, nn);
 
         myRecursiveTask task = new myRecursiveTask(a_n, b_n, nn);
         ForkJoinPool pool = new ForkJoinPool();
-        int[][] fastFJ = pool.invoke(task);
+        long[][] fastFJ = pool.invoke(task);
 
         return getSubmatrix(fastFJ, a.length, b[0].length);
     }
@@ -255,7 +255,7 @@ public class MatrixMultiplication {
      * Single-threaded matrix multiplication
      * algorithm by Strassen
      * */
-    private static int[][] multiStrassen(int[][] a, int[][] b, int n) {
+    private static long[][] multiStrassen(long[][] a, long[][] b, int n) {
         if (n <= 128) {
             return multiplyTransposed(a, b);
         }
@@ -263,31 +263,31 @@ public class MatrixMultiplication {
         n = n >> 1;
         ArrayList<Object> objects = new ArrayList<>();
 
-        int[][] a11 = new int[n][n];
-        int[][] a12 = new int[n][n];
-        int[][] a21 = new int[n][n];
-        int[][] a22 = new int[n][n];
+        long[][] a11 = new long[n][n];
+        long[][] a12 = new long[n][n];
+        long[][] a21 = new long[n][n];
+        long[][] a22 = new long[n][n];
 
-        int[][] b11 = new int[n][n];
-        int[][] b12 = new int[n][n];
-        int[][] b21 = new int[n][n];
-        int[][] b22 = new int[n][n];
+        long[][] b11 = new long[n][n];
+        long[][] b12 = new long[n][n];
+        long[][] b21 = new long[n][n];
+        long[][] b22 = new long[n][n];
 
         splitMatrix(a, a11, a12, a21, a22);
         splitMatrix(b, b11, b12, b21, b22);
 
-        int[][] p1 = multiStrassen(summation(a11, a22), summation(b11, b22), n);
-        int[][] p2 = multiStrassen(summation(a21, a22), b11, n);
-        int[][] p3 = multiStrassen(a11, subtraction(b12, b22), n);
-        int[][] p4 = multiStrassen(a22, subtraction(b21, b11), n);
-        int[][] p5 = multiStrassen(summation(a11, a12), b22, n);
-        int[][] p6 = multiStrassen(subtraction(a21, a11), summation(b11, b12), n);
-        int[][] p7 = multiStrassen(subtraction(a12, a22), summation(b21, b22), n);
+        long[][] p1 = multiStrassen(summation(a11, a22), summation(b11, b22), n);
+        long[][] p2 = multiStrassen(summation(a21, a22), b11, n);
+        long[][] p3 = multiStrassen(a11, subtraction(b12, b22), n);
+        long[][] p4 = multiStrassen(a22, subtraction(b21, b11), n);
+        long[][] p5 = multiStrassen(summation(a11, a12), b22, n);
+        long[][] p6 = multiStrassen(subtraction(a21, a11), summation(b11, b12), n);
+        long[][] p7 = multiStrassen(subtraction(a12, a22), summation(b21, b22), n);
 
-        int[][] c11 = summation(summation(p1, p4), subtraction(p7, p5));
-        int[][] c12 = summation(p3, p5);
-        int[][] c21 = summation(p2, p4);
-        int[][] c22 = summation(subtraction(p1, p2), summation(p3, p6));
+        long[][] c11 = summation(summation(p1, p4), subtraction(p7, p5));
+        long[][] c12 = summation(p3, p5);
+        long[][] c21 = summation(p2, p4);
+        long[][] c22 = summation(subtraction(p1, p2), summation(p3, p6));
 
         return collectMatrix(c11, c12, c21, c22);
     }
@@ -305,14 +305,14 @@ public class MatrixMultiplication {
 
     //******************************************************************************************
 
-    private static int getNewDimension(int[][] a, int[][] b) {
+    private static int getNewDimension(long[][] a, long[][] b) {
         return 1 << log2(Collections.max(Arrays.asList(a.length, a[0].length, b[0].length)));
     }
 
     //******************************************************************************************
 
-    static int[][] randomMatrix(int m, int n) {
-        int[][] a = new int[m][n];
+    static long[][] randomMatrix(int m, int n) {
+        long[][] a = new long[m][n];
         for (int i = 0; i < m; i++) {
             for (int j = 0; j < n; j++) {
                 a[i][j] = new Random().nextInt(100);
@@ -323,7 +323,7 @@ public class MatrixMultiplication {
 
     //******************************************************************************************
 
-    static void printMatrix(int[][] a) {
+    static void printMatrix(long[][] a) {
         for (int i = 0; i < a[0].length; i++) {
             System.out.print("-------");
         }
@@ -346,14 +346,14 @@ public class MatrixMultiplication {
 
     public static void test(int n, int m, int l) {
 
-        int[][] a = randomMatrix(n, l);
-        int[][] b = randomMatrix(l, m);
+        long[][] a = randomMatrix(n, l);
+        long[][] b = randomMatrix(l, m);
         long start, end;
 
         //****************************************
         //	TEST 1
         start = System.currentTimeMillis();
-        int[][] matrixByStrassenFJ = multiStrassenForkJoin(a, b);
+        long[][] matrixByStrassenFJ = multiStrassenForkJoin(a, b);
         end = System.currentTimeMillis();
         System.out.printf("Strassen Fork-Join Multiply [A:%dx%d; B:%dx%d]: \tElapsed: %dms\n", n, l, l, m, end - start);
         //****************************************
@@ -363,11 +363,11 @@ public class MatrixMultiplication {
         start = System.currentTimeMillis();
         int nn = getNewDimension(a, b);
 
-        int[][] a_n = addition2SquareMatrix(a, nn);
-        int[][] b_n = addition2SquareMatrix(b, nn);
+        long[][] a_n = addition2SquareMatrix(a, nn);
+        long[][] b_n = addition2SquareMatrix(b, nn);
 
-        int[][] temp = multiStrassen(a_n, b_n, nn);
-        int[][] matrixByStrassen = getSubmatrix(temp, n, m);
+        long[][] temp = multiStrassen(a_n, b_n, nn);
+        long[][] matrixByStrassen = getSubmatrix(temp, n, m);
         end = System.currentTimeMillis();
         System.out.printf("Strassen Multiply [A:%dx%d; B:%dx%d]: \tElapsed: %dms\n", n, l, l, m, end - start);
         //****************************************
@@ -375,7 +375,7 @@ public class MatrixMultiplication {
         //****************************************
         //	TEST 3
         start = System.currentTimeMillis();
-        int[][] matrixByUsual = multiply(a, b);
+        long[][] matrixByUsual = multiply(a, b);
         end = System.currentTimeMillis();
         System.out.printf("Usual Multiply [A:%dx%d; B:%dx%d]: \tElapsed: %dms\n", n, l, l, m, end - start);
         //****************************************
@@ -383,7 +383,7 @@ public class MatrixMultiplication {
         //****************************************
         //	TEST 4
         start = System.currentTimeMillis();
-        int[][] matrixByUsualTransposed = multiplyTransposed(a, b);
+        long[][] matrixByUsualTransposed = multiplyTransposed(a, b);
         end = System.currentTimeMillis();
         System.out.printf("Usual Multiply Transposed [A:%dx%d; B:%dx%d]: \tElapsed: %dms\n", n, l, l, m, end - start);
         //****************************************
@@ -397,19 +397,19 @@ public class MatrixMultiplication {
     //******************************************************************************************
 
     private static class Multipliers {
-        private final int[][] matrixA;
-        private final int[][] matrixB;
+        private final long[][] matrixA;
+        private final long[][] matrixB;
 
-        public Multipliers(int[][] a, int[][] b) {
+        public Multipliers(long[][] a, long[][] b) {
             matrixA = a;
             matrixB = b;
         }
 
-        public int[][] getMatrixB() {
+        public long[][] getMatrixB() {
             return matrixB;
         }
 
-        public int[][] getMatrixA() {
+        public long[][] getMatrixA() {
             return matrixA;
         }
     }
@@ -448,8 +448,8 @@ public class MatrixMultiplication {
             throw new IllegalArgumentException("The number of columns of the matrix A is not equal to the number of rows of the matrix B.");
         }
 
-        int[][] a = new int[rowsA][columnsA];
-        int[][] b = new int[rowsB][columnsB];
+        long[][] a = new long[rowsA][columnsA];
+        long[][] b = new long[rowsB][columnsB];
 
         int k = 4;
 
@@ -495,8 +495,8 @@ public class MatrixMultiplication {
     public static void main(String[] args) {
         Multipliers multipliers = validation(args);
 
-        int[][] matrixByStrassenFJ = multiStrassenForkJoin(multipliers.getMatrixA(), multipliers.getMatrixB());
-        int[][] matrixByUsual = multiply(multipliers.getMatrixA(), multipliers.getMatrixB());
+        long[][] matrixByStrassenFJ = multiStrassenForkJoin(multipliers.getMatrixA(), multipliers.getMatrixB());
+        long[][] matrixByUsual = multiply(multipliers.getMatrixA(), multipliers.getMatrixB());
 
         printMatrix(matrixByStrassenFJ);
         printMatrix(matrixByUsual);
@@ -505,4 +505,3 @@ public class MatrixMultiplication {
     }
 
 }
-
